requirements.txt
requests
pandas
duckdb
python-dotenv
